+++ rust/lib/src/lib.rs
@@ -1,0 +1,9 @@
+pub mod core_api_bindings;
+pub mod execution;
+pub mod networks;
+pub mod services;
+pub mod testsuite;
+
+pub fn say_hello() {
+    println!("Hello, world!");
+}